+++ geotrek/api/v2/views/tourism.py
@@ -86,8 +86,8 @@ class TouristicEventViewSet(api_viewsets.GeotrekGeometricViewset):
         api_filters.NearbyContentFilter,
         api_filters.UpdateOrCreateDateFilter,
     )
-    filterset_fields = ["bookable"]  # TODO find a way to overide automatic Swagger doc to add a description for this filter
     serializer_class = api_serializers.TouristicEventSerializer
+    filterset_fields = ["cancelled", "bookable"]
 
     def get_queryset(self):
         activate(self.request.GET.get('language'))