+++ docs/changelog.rst
@@ -13,6 +13,12 @@ CHANGELOG
 
 - Fix TouristicEvent with no end dates are not returned in APIv2
 
+**New features**
+
+- Add `cancelled` and `cancellation_reason` to TouristicEvents
+- Add `bookable` and `cancellation_reason` to TouristicEvents
+
+
 2.87.2 (2022-09-23)
 -----------------------